do not use a custom RunnableSpec
